godam-validator/requirements.txt
pandas
openpyxl
psycopg2-binary
python-dateutil
